render chapters differently for mac and iOS to account for platform differences
